General update (not recommended, I literaly forgot to commit at each new change)

+++ Models/Genre.cs
@@ -4,10 +4,9 @@ namespace Vidly.Models
 {
     public class Genre
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
         [Required]
         [StringLength(100)]
         public string Name { get; set; }
-
     }
 }
